pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "memory-bridge"
version = "0.1.0"
description = "Middleware for Agent LLM memory and context consistency"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "mem0ai>=2.0.2",
    "httpx>=0.28",
    "pydantic-settings>=2.7",
    "python-dotenv>=1.0",
    "qdrant-client>=1.13",
]

[project.scripts]
host-manager = "memory_bridge.host_manager:main"

[project.optional-dependencies]
dev = [
    "mypy>=1.15",
    "ruff>=0.11",
    "black>=25.1",
    "pytest>=8.3",
    "pytest-asyncio>=0.25",
    "httpx>=0.28",
]

[tool.hatch.build.targets.wheel]
packages = ["src/memory_bridge"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.ruff.lint.isort]
known-first-party = ["memory_bridge"]

[tool.mypy]
strict = true
python_version = "3.11"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
